+++ apps/cli/src/commands/variable/create.variable.ts
@@ -41,11 +41,22 @@ export default class CreateVariable extends BaseCommand {
       },
       {
         short: '-e',
-        long: '--entries [entries...]',
+        long: '--entry [entries...]',
         description:
-          'An array of key-value pair (value and environmentSlug) for the variable.'
+          'An array of values for the variable. If specified, should be in the form <environment slug>:<value>'
       }
     ]
+  }
+
+  getUsage(): string {
+    return `keyshade variable create <project slug> [options]
+    
+  Create a variable
+  keyshade variable create project-1 --name "PORT" --entry "alpha=8080"
+  
+  Create a variable with note
+  keyshade variable create project-1 --name "PORT" --note "This is a variable"
+    `
   }
 
   canMakeHttpRequests(): boolean {
@@ -110,13 +121,16 @@ export default class CreateVariable extends BaseCommand {
     }
 
     const parsedEntries = entries.map((entry) => {
-      const [environmentSlug, value] = entry.split('=').map((s) => s.trim())
-      if (!environmentSlug || !value) {
-        throw new Error(
-          `Invalid entry format: ${entry}. Expected format: "environmentSlug=value"`
-        )
+      const entryObj: { value: string; environmentSlug: string } = {
+        value: '',
+        environmentSlug: ''
       }
-      return { environmentSlug, value }
+      entry.split(' ').forEach((pair) => {
+        const [key, value] = pair.split('=')
+        entryObj.environmentSlug = key
+        entryObj.value = value
+      })
+      return entryObj
     })
 
     return {