+++ apps/cli/CHANGELOG.md
@@ -82,3 +82,8 @@
 - Formatted outputs of `keyshade secret list` and `keyshade secret revisions` commands
 - Update `keyshade secret rollback` to check for missing environment or version
 - Added usage examples in secret commands
+- Removed `keyshade variable get` command
+- Updated `keyshade variable update` and `keyshade variable create` commands to include `-e/--entry` flag for specifying values to environments
+- Formatted outputs of `keyshade variable list` and `keyshade variable revisions` commands
+- Update `keyshade variable rollback` to check for missing environment or version
+- Added usage examples in variable commands